# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HowMuch API Documentation",
        "description": "This is the API documentation for the HowMuch API.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/invoices/{id}": {
            "get": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Get an invoice",
                "operationId": "5d25d392f44d2cf516114d871e72ecb7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Invoice ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "invoice": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "example": 2
                                                        },
                                                        "date": {
                                                            "type": "string",
                                                            "example": "2022-06-28"
                                                        },
                                                        "total": {
                                                            "type": "string",
                                                            "example": 100
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "purchases": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "date": {
                                                                "type": "string",
                                                                "example": "2022-06-28"
                                                            },
                                                            "items": {
                                                                "type": "array",
                                                                "items": {
                                                                    "properties": {
                                                                        "id": {
                                                                            "type": "string",
                                                                            "example": 1
                                                                        },
                                                                        "title": {
                                                                            "type": "string",
                                                                            "example": "Compra no ifood"
                                                                        },
                                                                        "value": {
                                                                            "type": "string",
                                                                            "example": 90.9
                                                                        },
                                                                        "type": {
                                                                            "type": "string",
                                                                            "example": "food"
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/invoices/{id}/chart": {
            "get": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Get the invoice chart",
                "operationId": "7fddb687f1f183fd340eaada5165c847",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Invoice ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "type": {
                                                "type": "string",
                                                "example": "food"
                                            },
                                            "value": {
                                                "type": "string",
                                                "example": 100
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/invoices/current-invoice": {
            "get": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Get the current invoice",
                "operationId": "98b8bd671d482bc87993c5b5724e0749",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "invoice": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "example": 2
                                                        },
                                                        "date": {
                                                            "type": "string",
                                                            "example": "2022-06-28"
                                                        },
                                                        "total": {
                                                            "type": "string",
                                                            "example": 100
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "purchases": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "date": {
                                                                "type": "string",
                                                                "example": "2022-06-28"
                                                            },
                                                            "items": {
                                                                "type": "array",
                                                                "items": {
                                                                    "properties": {
                                                                        "id": {
                                                                            "type": "string",
                                                                            "example": 1
                                                                        },
                                                                        "title": {
                                                                            "type": "string",
                                                                            "example": "Compra no ifood"
                                                                        },
                                                                        "value": {
                                                                            "type": "string",
                                                                            "example": 90.9
                                                                        },
                                                                        "type": {
                                                                            "type": "string",
                                                                            "example": "food"
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/invoices/next-invoices": {
            "get": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Get the next invoices",
                "operationId": "1070b9497820caec928ead0cdf3c51e1",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "total": {
                                                    "type": "string",
                                                    "example": 100
                                                },
                                                "invoices": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "string",
                                                                "example": 1
                                                            },
                                                            "date": {
                                                                "type": "string",
                                                                "example": "2022-06-28"
                                                            },
                                                            "total": {
                                                                "type": "string",
                                                                "example": 100
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}